Fix elf_begin() failure on i386
On i386 with systems that have more than 4 GiB physical memory, the vmcore is
larger than 4 GiB, and elf_begin() fails on that systems because the libelf
tries to map the whole dump into memory. However, we only need the beginning
of the ELF file, and therefore, we use mmap() manually and use the elf_memory()
interface.