Fix passing "sysrq=1" to the kdump kernel
The kdump init script adds "sysrq=1" to the kdump kernel command line
unconditionally. This is highly confusing if the user wants to debug
the kdump kernel and adds "debug" to the command line, because the
"debug" setting gets overridden later from the kdump initrd.

Use sysrq=yes because that actually was the intention: to enable sysrq
in kdump kernel.

References: bnc#609143
Signed-off-by: Petr Tesarik <ptesarik@suse.cz>