Pass acpi_rdsp boot option to kdump
EFI cannot be initialized twice, so if the primary kernel is booted
in EFI mode, the secondary (kdump) kernel must not try to initialize
it again. However, without EFI tables, the secondary kernel may not
be able to find the ACPI tables and will fail anyway. So, the address
of the RSDP table must be passed too.

References: bnc#717263
References: https://bugzilla.redhat.com/show_bug.cgi?id=681796
Signed-off-by: Tony Jones <tonyj@suse.de>
Acked-by: Petr Tesarik <ptesarik@suse.cz>