Let mkinitrd decide static/dynamic if KDUMP_NETCONFIG=auto
If KDUMP_NETCONFIG is set to "auto", mkdumprd always passes "-D auto" to
mkinitrd. This does not work for static IP configurations. If we pass no
networking arguments at all, mkinitrd decides automatically.

In SLE11 SP1, this happened to work, because mkinitrd did not honour the
interface type as it should, but it's broken for SP2.

References: bnc#706059
Signed-off-by: Petr Tesarik <ptesarik@suse.cz>