Recognize compressed ELF as a valid kernel file
While IBM fixed the case for uncompressed ELF files, there is still an
issue with compressed ELF files.

The trouble is that gzdopen() does not rewind the file, so if you call
it repeatedly on the same file, then the second call will not find the
compressed header and assume that the file is uncompressed. The call to
gzseek() comes too late, the header is read without decompression, and
the file is not recognized as an ELF file.

To fix both cases (compressed and uncompressed), the file pointer must
be set to zero _before_ calling gzdopen().

References: bnc#732768
Signed-off-by: Petr Tesarik <ptesarik@suse.cz>