Copy makedumpfile-R.pl to the kdump initrd
Because makedumpfile is now linked dynamically against a library under
/usr, it can't go to /sbin, so it goes to /usr/sbin. In fact, all
makedumpfile files are now installed in /usr/sbin, but makedumpfile-R.pl
cannot be found in that location during dumping, so kdumptool fails.

It is wrong to take it from the mounted system, in the first place. If it
must be copied to the target directory, it should be included in the initrd.

References: bnc#734468
Signed-off-by: Petr Tesarik <ptesarik@suse.cz>