Add "none" dump format
Now that we have the ability to dump the dmesg log after a crash, it's
useful for users who don't have any use for a crash dump other than
to generate the dmesg log to avoid saving the dump at all. This avoids
both using the disk space required for a lot as well as the time
involved in waiting for the dump to complete only to later have to
remove it manually.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: Petr Tesarik <ptesarik@suse.cz>