Avoid one forgotten '\r' on dumb terminals
This changeset modifies the logic slightly.

TerminalProgress::progressed() now always ensures that output starts at
the beginning of a line. For dumb terminals it emits a newline, otherwise
it uses a carriage return. After the progress is written, the cursor stays
at the end of the line.

This doesn't work very well with TerminalProgress::stop(), because
clearLine() does nothing on dumb terminals, so we must explicitly add a
newline before the last line.