Do not set up iommu pass-through for the kdump kernel
The crash kernel will not find the megaraid root (for some unknown reason)
if iommu=pt is specified on the crash kernel command line. While it would
be better to fix the real problem, the crash kernel doesn't care if there
is a pass-through domain, so it can be always disabled.

Fixes bnc#804800.