fadump: avoid duplicate writes to /sys/kernel/fadump_registered
IBM engineer checked the source code, and found that in the kernel, the
fadump_registered store method doesn't allow the duplicated assignments.

I personally find it silly to treat such a write as an error (and even the
following workaround contains a race window), but it's the easiest fast fix
for SLES 11 SP3.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>
References: bnc#814780

---
 init/boot.kdump |   12 ++++++++++--
 1 file changed, 10 insertions(+), 2 deletions(-)