Include hpwdt in kdump initrd if the hardware is present
When a crash is induced with echo c > /proc/sysrq to start a dump, the hpwdt
driver is not involved and does not turn off the timer.   Therefore the ASR
timer is still running as the dump proceds and when the timeout expires it
issues NMI and system reset that terminates the dump.

If the hpwdt driver is loaded in the crashkernel, it can turn off the HW timer,
and thus prevent the ASR timeout from occurring.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>
References: bnc#807873

---
 init/setup-kdump.sh |   42 ++++++++++++++++++++++++++++++++++++++++++
 1 file changed, 42 insertions(+)