Add KDUMP_FADUMP as a configuration option
The FADUMP option was previously controlled by the kernel command line.
This is suboptimal, because kdump can never say with certainty what is
going to be used on the next boot. It can also be obscure from an
administrator's point of view, because kdump configuration is not
centralized.

This changeset only adds the option, but doesn't use it for anything.