Add panic=1 to the kdump kernel command line
If the kdump kernel crashes, the default is to halt the machine. This
can be changed with the panic parameter, so add it to the kdump
command line. However, if a custom panic parameter already exists,
don't override it.

Rationale:

1. After a default install, kdump kernel will use "panic=1"
2. If the admin specifies a custom timeout in the bootloader, it will be used
   both for the primary kernel and the kdump kernel automatically.
3. If the admin wants to set different values for the primary kernel and the
   kdump kernel, they can put a "panic=" parameter to KDUMP_COMMANDLINE_APPEND.

The above should cover all possible scenarios, while providing a sensible
default.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>
References: bnc#809088