Fix an off-by-one bug in OptionParser::parse()
OK, this bug is not so obvious. The argv pointer will be passed to
getopt_long(), which assumes that the first argument (argv[0]) is
the command name. For subcommands, it should be subcommand name.

Note that the return value of parsePartial() cannot be added to the
index directly, because the subcommand name must be subtracted again.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>