Make a static library for build
There is no need to recompile every module for the unit tests. In fact,
the recompilation should be avoided to make sure exactly the same code
is tested which is also present in the binary.

With CMake, the only way to achieve this is to move all classes to
a static library which is used at build time (does not get installed).

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>