Fix subcommand initialization
The initialization magic no longer works after converting the common files
to a static library. Since there is no explicit reference to the object
files that contain global Subcommand objects, they are not included in the
link, and hence never initialized.

The objects must be initialized explicitly. But let's do it in main.cc,
which is not part of the library.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>