Remove unnecessary check for file existence in FileUtil::mkdir()
First of all, the check is racy. If the directory is created after the
stat() syscall, mkdir() can still fail.

Second, there is no need to check if the path exists, because this case
can be distinguished by looking at the errno value if mkdir() fails.

Third, this patch changes the behaviour for dangling symlinks. Previously,
it would be treated as non-existent, mkdir() would fail with EEXIST, and
the method would raise an exception.  After the change, a dangling symlink
is treated as success. Note that this cannot break the assumption that a
valid directory exists after the call to FileUtil::mkdir(), because this
assumption was broken already if the path referred to a non-directory.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>