Remove stripPrefix()
The implementation is either non-intuitive, or inefficient, or both.
Basically, stripPrefix() must check if the string starts with the given
prefix, but this information is also needed by the caller in most cases.

So, startsWith() gets called twice. Additionally, it is not clear if
the method should modify the object, or return a substring. Both is
inefficient at times.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>