Process: use streams for stdin, stdout, stderr
This cleans up the interface quite a bit. Admittedly, it also costs a bit
of efficiency (because all data must be copied by the parent process,
even if it could be passed directly), but there was not much use of it
in kdumptool, anyway.

This also improves test coverage in testio, as it now uses the same code
path as kdumptool.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>