Process: clear flags before attempting read
If the failbit is set in the input stream, then read() will never get any
data. At the same time, the stream is not bad(), and may not be at eof().
OTOH the fail bit is also set by read() on partial reads, which is an
expected condition at EOF, so it shouldn't be treated as an error.

To solve this situation, clear the state bits before reading.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>