Translate mount points between running and kdump
The thing is that the original mount point may not be available in the kdump
environment, e.g. given these mountpoints:

/
/something
/something/subdir/

The last mount point does not exist in the root filesystem, so it cannot be
mounted at that location.

The solution is to give each required filesystem its own mount point in the
kdump environment, translate the original mount point to /mnt$n and copy
the remaining path after that.

It is not possible to do without a prefix completely, because some directory
names may clash with the pre-existing directories in the initrd. I'm using
the /kdump prefix instead and put everything below there.

Note that two directories are special:

/kdump/root
  - corresponds to the root filesystem, which is already mounted by initrd

/kdump/boot
  - needed when copying the kernel image

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>