setup-kdumpfs: parse fstab only once
The previous behaviour was suboptimal. First, the kdump mount points were
retrieved from /etc/fstab. Next, each mount point was searched again in
fstab, this time remembering the device, fstype and options. A lot of the
code provided fallbacks if the mount point was not found in fstab, but
this case need not be considered at all.

The new algorithm is simple:

1. find the mount points (and remember all required fields)
2. write entries to fstab.kdump, while resolving each device to UUID

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>