Add dracut network activation to mkdumprd
Note the handling of KDUMP_NETCONFIG=auto. The network module is loaded
only if we know already that network is needed. The kernel option for
dracut is added even if no network requirements are found from mkdumprd.
That's because another dracut module may later drag in network as a
dependcy. However, network is set up only if it is required to mount the
root filesystem (not an additional mount).

By adding the option unconditionally the result is:

a. If network module is included in the initrd, it will set up network.
b. If network is not included, the option is ignored at boot time.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>

--HG--
extra : amend_source : fcf5f8fa82a0da43f6b69ac4683ba3b7a49d633a