Include original mountpoint path in kdump environment
The target path gets copied into notificiation mails, but the admin
has no idea which path was mounted on /mnt$i in the kdump environment,
and it is confusing.

It is still necessary to prefix the mount path with /mnt$i, because
the corresponding mount point may not exist if it was on a third
mount that is not included in the kdump environment.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>