SubProcess: Create pipes with O_CLOEXEC
This way the pipes are closed automatically on exec(), and only the
destination file descriptors are passed to the child process.

This change has implications when another child is spawned from the
same parent, as there would be no way for the user of SubProcess to
close the parent ends of the previous child's pipes in the forked
process, but leave them open in the parent.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>