Specify path to all SSH files based on home directory by default
The reason is that openssh does not use the HOME environment variable,
but gets the home directory using nss (i.e. from /etc/passwd). Since
it cannot be guaranteed that the default is available (e.g. mkinitrd
uses /root to mount the root file system), it is necessary to specify
all files explicitly.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>