calibrate: correctly align the size of memmap on s390
The s390(x) vmemmap contains a multiple of PAGES_PER_SECTION struct
pages, see arch/s390/kernel/setup.c.

PAGES_PER_SECTION is defined as SECTION_SIZE_BITS - PAGE_SHIFT:

- s390x has PAGES_PER_SECTION = 28
- s390 has PAGES_PER_SECTION = 25

Implementation note: The "maxpfn" variable is not the maximum PFN.
It is shifted right by 10, because all calculations are in KiB (it
is the number of "kibipages").

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>