Remove root= and resume= from the kdump kernel command line
Resume from hibernation should never be even attempted in kdump.
If using dracut, the module is even omitted from the initrd, so
passing the command line option only consumes command line space.

The default root device is not needed either, because all required
filesystems are handled by the mount dracut module. In fact, if the
root filesystem is also mounted, it may be mounted twice.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>