Dracut-systemd support
Adjust the dracut module to work with the systemd module.

A new service is added to the initrd. I could not use a pre-pivot
hook, because dracut-pre-pivot.service does not send the output
to the terminal, so dump progress was not visible.

It is also necessary to avoid mounting the root filesystem. All
filesystems needed by kdump are mounted under /kdump and set up
with the --mount dracut option.

The system root mount can be generated in two ways:

a. by initrd-fstab-generator (if "root=" is passed on the kernel
   command line)

b. by dracut-rootfs-generator (if the root option is stored in
   /etc/cmdline.d/*)

The root= parameter is already removed from the command line, but
it is always saved to /etc/cmdline.d/95root-dev.conf if dracut is
called with --hostonly-cmdline. This is the default in SLES (cf.
/etc/dracut.conf.d/01-dist.conf) and it is required to get correct
configuration for many setups.

So, the only way to avoid mounting /sysroot is to remove the
appropriate systemd generator from the initrd.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>