Get required multipath wwids from sysfs
The dracut code path must use for_each_host_dev_and_slaves_all to
find multipath devices that are "hidden" beneath e.g. an LVM volume.
That will iterate over all devices in the chain down to each physical
device. Note that /lib/udev/scsi_id sends a low-level SCSI command to
the device to find out the WWID. If it sends it to an unreachable
device in an active-passive multipath setup, it may hang indefinitely.

But the WWID is already available as the name of the multipath device,
so we can get it from the kernel via sysfs.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>