Always turn on fadump if KDUMP_FADUMP is true
The problem is that fadump-related files under /sys/kernel are
created only if the kernel is booted with fadump=on. So, this
is a chicken-and-egg problem, and a reboot is always necessary if
the current kernel is loaded without fadump=on.

However, if the command line is not modified here, the admin must
first add fadump=on to the bootloader command line, and then enable
KDUMP_FADUMP, i.e. enable it in two different places. If the command
line is modified unconditionally, there is only one "big switch" for
FADUMP.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>