Reuse kdump target mount point handling for /boot directory
Target mount points use a general algorithm to map paths in the
primary system to mount points in the initrd. The /boot directory
was handled separately, because the kdump_* arrays are also used
to modify the kdump config file.

Let's add a special "protocol" which should be handled like "file",
except it does not end up in KDUMP_SAVEDIR. This allows to reuse the
same mechanism to map it to an initrd path, and the root/boot special
cases can go away.

There's one small thing: kdumptool (in the initrd) does not know the
mapping. Originally, special symbolic links were created. We can
retain this logic by (ab-)using kdump_URL[] to pass the desired
target link name.

Note that the "root" symlink is no longer needed. It was only ever
used for the "boot" symlink if /boot was on the root filesystem.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>