Keep /sysroot mount point if fadump is on
As FADUMP uses the normal initrd, the system must be able to boot up
eventually, so normal mount points are required.

But mounting the same device twice may not be possible because of
conflicting mount options, so system root may not be mountable under
/kdump. In order to keep the kdump-specific mount point hierarchy,
convert any such mount to a bind mount from /sysroot.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>