Make sure system root is not mounted in kdump initrd
All required mounts are added under /kdump, including the root
filesystem (if needed at all). But initrd may try to mount the
system root filesystem nevertheless, which fails if it is also
mounted under /kdump.

This commit passes a special keyword for the root= parameter to
avoid the second mount. Originally, the root= parameter was omitted,
but that did not work on platforms with DeviceTree (e.g. ppc64),
because kexec-tools added this parameter automatically.

Note that the "root=" parameter may be interpreted by two different
pieces of software: either by the default systemd fstab-generator,
or by dracut's rootfs-generator. The root=kdump parameter takes
care of both, because:

1. anything that doesn't start with a slash ('/') is ignored by
   the systemd generator,
2. kdump-root.sh sets rootok=1, so dracut does not panic because
   of unknown root= parameter

References: bsc#900134
Signed-off-by: Petr Tesarik <ptesarik@suse.cz>