Start a SFTPPacket implementation
The sftp protocol is sufficiently flexible to allow saving makedumpfile's
output directly, but the sftp client is not (it has no way to copy standard
input to a target file). This is why kdumptool must implement the necessary
protocol commands itself.

This commit adds testing infrastructure and a dummy class for creating an
empty SFTP packet.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>