Clean up kdump mounts before continuing with FADUMP
Before pivoting the root, the initramfs content is wiped. This cannot be
performed if there are still active mounts under /kdump, resulting in
kernel panic. If the system is correctly configured for FADUMP, it will
try to save a new dump from the panic after reboot, which fails the
same way and so on...

References: bsc#923790
Signed-off-by: Petr Tesarik <ptesarik@suse.cz>