Remove unnecessary dependencies from kdump-save.service
The attempt to serialize with other parts of dracut is wrong,
because the dracut pieces are linked together using Wants=, but
kdump-service has a hard requirement on the kdump mount points,
so it may run out of sync anyway.

The dependency on dracut-pre-pivot.service in fact creates a cycle,
because dracut-pre-pivot wants to be run after initrd.target, but
initrd.target wants kdump-save.service.

References: bsc#942895
Signed-off-by: Petr Tesarik <ptesarik@suse.com>