Reload kdump only once if multiple udev events happen in parallel
The kexec(2) system call is not re-entrant, so kdump reload may fail
if another reload is already in progress. Solve this by serializing
the requests using a lock file under /var/run.

Note that it is sufficient to re-run kexec(1) once after all memory
configuration changes are finished. To avoid unnecessary runs, the
following logic is implemented: If a reload is already in progress,
all subsequent reloads exit immediately, letting the running instance
redo the job. A jobs file is used to tell the running instance that
there are more changes after the first one, so it must reload the
kdump kernel one more time.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>