Add a network check with configurable timeout
Before accessing a remote host, check that the network layer is
ready. The tool will stop waiting if network status does not change
within a configurable timeout (KDUMP_NET_TIMEOUT), but it will try
to save it anyway, because the underlying helpers might have an
alternative way to connect to the remote host.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>