calibrate: Add memory for SWIOTLB if maxpfn is above 4G
The kernel may need a SWIOTLB if a system has RAM above 4G. Note that
a system may have less than 4G total RAM but still physical addresses
above 4G (because of memory holes). Consequently, the code must check
/proc/iomem rather than just memtotal.

Of course, if the system has a hardware IOMMU, software IO TLB may
not be needed. There is no (straightforward) way to check on a
running system if SWIOTLB is used or not, so let's always reserve
the memory for it, except when running on a system with all RAM
below 4G. This is acceptable, because only such small systems may
be concerned about reducing available RAM by 64M.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>