Get configured interface kind using wicked
The kdump interface may not be up or even present when running
mkdumprd, so ip(8) cannot be used. To solve it, read the configured
values using wicked.

The problem with wicked show-config is that the interface kind is not
contained in the XML; it must be inferred from the presence or absence
of certain elements... Oh, well.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>