Handle dump files with many program headers
To reduce memory requirements, kdumptool reads only the beginning of
the dump file. If there are too many program headers, this size is
insufficient, and the program header table appears truncated.

Unfortunately, gelf_getehdr() cannot be used, because it fails when
program headers are truncated. Raw ELF headers must be examined.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>