Remove 'notsc' from the kdump kernel command line
The kdump kernel boots on the crashed CPU, which may not be the BSP.
Boot may then fail, because the early code switches to legacy interrupt
delivery, which may fail on AP for many reasons. See also
http://thread.gmane.org/gmane.linux.kernel/607252

I don't see any reason to disable TSC in the kdump kernel. People may
want to disable TSC because in some cases it can cause time to be
unreliable. This should be really rare with a single CPU boot and more
importantly, we don't care in the kdump environment.

Signed-off-by: Jiri Bohac <jbohac@suse.com>
Acked-by: Petr Tesarik <ptesarik@suse.com>