Determine network configuration in dracut's check() function
To allow using the variables set by kdump_ifname_config() from other
functions than the caller, they cannot not be set in cmdline(),
because this function is called by dracut in a sub-shell.

Instead, parse the network configuration during check() and use
global variables to pass the configuration to cmdline().

Signed-off-by: Petr Tesarik <ptesarik@suse.com>