Include required kernel modules for inactive interfaces
Dracut's network module already includes modules for bridges, bonding
and vlan, but if dracut runs with the --hostonly option, they are
included only if the corresponding module is loaded. This logic is
wrong if kdump uses a non-default interface which is not started
automatically (e.g. with STARTMODE='manual').

Make sure that all necessary kernel modules are included by calling
instmods with hostonly explicitly set to an empty string.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>