Rename the systemd services to make them less confusing
kdump-early.service loads the kernel and initrd as soon as possible.
kdump.service rebuild the initrd if it is neccessary and load it.
Otherwise it dose not reload anything. Now the status of kdump.service
reflects the real status of kdump loading which is less confusing.

Of course, kdump may still use an outdated initrd if the system crashes
between kdump-early.service and kdump.service, but better than nothing.