Convert sysroot to a bind mount in kdump initrd
In SLES 12 SP2, systemd-fstab-generator no longer ignores non-device
root mounts, so it tries to run an actual mount command for root=kdump.
This fails, of course, because "kdump" is not mountable.

To solve this, pass "rootflags=bind" to the panic kernel, so systemd
can create a (bogus) bind mount and be happy.

See also kdump-root-parameter.patch.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>