Fix kdump-related services
kdump.service should be the "big switch" to turn kdump on and off. We
can achieve this by adding an Also= directive.

kdump-early.service is merely a way to load a kdump kernel a early
during the boot process if and only if the kdumprd already exists (e.g.
from a previous boot). There may be some use cases for enabling only
the early service without kdump.service.

OTOH, it is not an error if the initrd is not found during the early
stages. This is indicated by exit code 6 (originally conforming to LSB
"program not configured" exit code).

References: bsc#1021484
Signed-off-by: Michal Koutny <mkoutny@suse.com>
Acked-by: Petr Tesarik <ptesarik@suse.com>