Do not free fadump memory when reboot is requested
Freeing fadump memory can take a long time and doing it when we are going to
reboot anyway is pointless.

In bsc#1034169 a kernel oops is triggered by freeing fadump memory killing
save_dump.sh and preventing the reboot requested by the user from happening
entirely.

Avoid the issue and speed up reboot by not freeing fadump memory when reboot is
requested.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>