Fix save_dump to NFS targets
For NFS, the parent directory of the target must be mounted, because
the target directory itself does not exist yet.

For this to work, split the target path to directory (i.e. parent
directory) and basename. Then construct a unique mount point from the
host name and the parent directory, and finally append the base name
again to form a subdirectory in the target (to be created by the file
transfer method).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>