Restore only static routes in kdump initrd
All existing routes are now added through the rd.route= dracut
parameter. However, this includes routes that need not or should not
be added explicitly (e.g. installed automatically by the kernel or
dynamically by a routing daemon).

If possible, use wicked to get configured routes. If not, use the
routing protocol identifier to limit the list.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>