Recover from missing CRASHTIME= in VMCOREINFO
CRASHTIME= may be missing in Xen Dom0 dumps.
Vmcoreinfo::getLLongValue throws an exception in that case, but then
OSRELEASE will not even be attempted. Consequently, kernel and
system map files are not copied, producing the following message:

INFO: Don't copy the kernel and System.map because of missing crash kernel release.

References: bsc#1112387
Signed-off-by: Petr Tesarik <ptesarik@suse.com>