Preserve white space when removing kernel command line options
The function was originally designed to remove unwanted options
from the panic kernel command line. It is now also used to check
whether the current command line contains a fadump option or not,
but the check is broken, because remove_from_commandline() may
change the amount (and type) of white space. For example, it
always adds an extra space if the original string ends with a
space.

Modify the function to copy original separators verbatim, except
that any white space preceding the removed option is removed.

Fixes: a4718a2c7b714d0594903fc8dc5ae47252a9b9ba
References: bsc#1117652
Signed-off-by: Petr Tesarik <ptesarik@suse.com>