remove console=hvc0 from commandline
If kdump is prepared in a dom0, the dom0 kernel cmdline is used
verbatim for the kdump kernel, modulo known-to-be-bad options.
As a result 'console=hvc0' is passed to the kdump kernel.
In a native environment this option causes the kdump kernel to hang,
the system never comes back.

Remove it from the kdump cmdline to make sure the kdump kernel boots.

Signed-off-by: Olaf Hering <olaf@aepfle.de>