calibrate: Fix kernel command line parsing
A kernel parameter value can be quoted (param="val"). This quoting
is not removed properly. It is necessary to follow the logic of
kernel's next_arg() in lib/cmdline.c more closely.

Note that if the value as well as the whole argument are quoted,
the Linux kernel removes only one trailing quote. For instance:

  "root="/dev/sda1""

is parsed as /dev/sda1" (with the trailing quote). I consider this
a bug in the kernel option parser and I'm not redoing it here.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>