Reimplement Stringutil::stringv2charv()
The old approach is inherently wrong:

- This is a static method operating on a StringVector using no
  data from its class, so essentially a function in a namespace.

- Allocated data does not follow the RAII principle: It is
  expected that Util::freev() is always called on the result, but
  if an exception is raised, this may never happen.

In an ideal world, we could simply keep a vector of char* pointers
returned by c_str(). Unfortunately, c_str() returns a pointer to a
const char, but execvp() requires an array of constant pointers to
a (mutable) char. I doubt it actually modifies the strings in that
array, but---better safe than sorry.

To get mutable strings, a copy of each string in the vector must
be allocated, but we can't simply keep the pointers in a vector of
char*, because the pointers would not be owned by anybody.
Instead, we need a vector of std::unique_ptr, but since the
in-memory representation of a unique_ptr may not match that of the
owned pointer type, a second vector must be allocated for the
return value.

The only good side effect is that we can return a `char**` and not
a `char *const *`, because the pointers in the return vector are
not used for de-allocation.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>