BufferDataProvider: Use memory pointers instead of ByteVector
Since all users can ensure that a pointer to the buffer memory
stays valid throughout the lifetime of the BufferDataProvider
instance, we can avoid copying.

Note: It would be probably safer to use a smart pointer, but then
we could not use std::string::c_str(). I'm leaving everything as
is for now.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>