Reimplement SubProcess::PipeInfo with a class hierarchy
Just look at the code in SubProcess::spawn(), why it's necessary
to catch exceptions, clean up and pass on, and you know that this
is the wrong way to do it.

The new method is still ugly (yes, that typeid() hurts), but it's
done to maintain some backward compatibility and keep a reasonable
amount of changes in each commit.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>