Fix the testprocess test case
The test case should not close the parent end of the pipe, because
the fd is still stored in the SubProcessFD object. The destructor
will try to close the fd on shutdown, but by that time the fd
number may have been reused...

Of course, the file descriptor must be closed to signal EOF to the
subprocess that reads from stdin, but by using the close() method,
the file descriptor stored in the SubProcessFD is correctly
uninitialized.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>