Convert all easy cases to the new SubProcess API
The conversion of most old API users is straightforward.

The testprocess program is a bit more complex. No exceptions are
raised with the new API, because a nullptr can be returned
instead. The tests for the correct type are a bit pointless now,
but at least it checks the API calls themselves.

This leaves ProcessFilter::IO and descendants, but I'm moving that
to a separate commit to avoid cluttering this one.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>