Rewrite fadump-init in plain C
The fadump-init.sh shell script runs as the very first process at
boot, before even systemd. It is quite difficult to achieve decent
error handling in bash, because it may rely on features that are
not yet available in this environment.

Additionally, the goal cannot be achieved without running external
processes, but all this forking and executing takes some time,
which is undesirable at this early stage.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>