init-fadump: Move /init.dracut over /init
The actual init may assume that it is running as /init, so rename
the binary to the expected name before executing it.

If that fails, fall back to executing it under the modified name,
because that's very likely to work and is probably better than
exiting init and causing a kernel panic.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>