init-fadump: Do not mount /newroot
Since initramfs runs from rootfs which cannot be unmounted anyway,
moving to another mounted memory filesystem does not bring much
benefit, but as any unnecessary code, it may fail.

Originally, the added benefit was that switch_root empties rootfs
content, freeing up associated memory. Since this utility is now
reimplemented in init-fadump, the recursive remove can be modified
to keep /fadumproot.

The sequence of mount operations may look a bit convoluted. To
replacethe root of the filesystem, a move operation (MS_MOVE) must
be done. The source of the move operation must be a mount point,
however, and /fadumproot is converted to one by bind-mounting it
over itself.

The nice thing about this commit is that the dependency on /bin/mv
can be dropped, making the binary self-contained.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>